requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
numpy==1.26.4
python-dotenv==1.0.1
google-generativeai==0.3.2
